- Implemented purchase_order - refractored payment gateway

+++ jft-market/src/main/java/com/jft/market/api/ProductBean.java
@@ -11,11 +11,11 @@ public class ProductBean {
 
 	private Long id;
 	private String name;
-	private Integer price;
+	private Long price;
 	private String description;
 	private String features;
 
-	public ProductBean(Long id, String name, Integer price, String description, String features) {
+	public ProductBean(Long id, String name, Long price, String description, String features) {
 		this.id = id;
 		this.name = name;
 		this.price = price;
